внешние доработки, избавление от ошибок

+++ src/redux/actionCreators/postsActionCreators.js
@@ -13,7 +13,7 @@ const getPostsFromServer = (postsFromServer) => ({
 })
 
 // получение всех постов с сервера
-export const getPostsFromServerQuery = (page = '', limit = '', filter = '', setLoading) => async (dispatch) => {
+export const getPostsFromServerQuery = (setLoading, page = '', limit = '', filter = '') => async (dispatch) => {
   const response = await axiosInstance.get(
     `posts/paginate?page=${page}&limit=${limit}&query=${filter}`,
   )
@@ -54,11 +54,13 @@ const deletePost = (id) => ({
 // удаление поста по id
 export const deletePostQuery = (id) => async (dispatch) => {
   try {
-    await axiosInstance.delete(
-      `posts/${id}`,
-    )
-
-    dispatch(deletePost(id))
+    const isDelete = confirm('Точно хотите удалить пост?')
+    if (isDelete) {
+      await axiosInstance.delete(
+        `posts/${id}`,
+      )
+      dispatch(deletePost(id))
+    }
   } catch (e) {
     const codeError = e.message.slice(-3)
     if (codeError === '403') {
